Enable overflow checking on the release build
This is probably best for smart contracts, where math errors are likely
to be security vulnerabilities, and so overflow should be caught at
run-time.  Ideally, this could be done without enabling all debug
assertions, but Rust doesn’t provide for this.

The fold in src/bridge.rs has had overflow checking added explicitly.